ZeroProject/requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-rest==0.8.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
jsonfield==3.1.0
lxml==4.6.3
Pillow==8.3.1
platformdirs==2.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.7.0
webdriver-manager==3.4.2
whitenoise==5.3.0
